requirements.txt
progress
opencv-python
Pillow
torch
numpy
tensorflow-gpu
matplotlib